requirements.txt
streamlit
google-generativeai
langchain
langchain-community
faiss-cpu
sentence-transformers
punctuators
git+https://github.com/jianfch/stable-ts.git